changed various SoyEvents to std::functions to remove use of locks

+++ src/SoyEvent.cpp
@@ -1,7 +1,7 @@
 #include "SoyEvent.h"
 
 
-std::atomic<int> SoyListenerId::mListenerCounter(1);
+std::atomic<uint64_t> SoyListenerId::mListenerCounter(1);
 
 #if defined(UNITTESTCPP_H)
